+++ CHANGELOG.md
@@ -7,8 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- Added first draft of the README
 - The ability to run a user's own custom shell scripts placed in `./scripts/custom-command.sh`
+- Discord Integration. The user now has the ability to configure settings to post messages into their Discord server.
+  - Unhandled exceptions. When the system runs into an unhandled error it will now post a message to your configured Discord server.
+- Password input is now hidden during setup
+- Command Terminal. You can now open a command terminal with '-c' with no additional arguments.
 
 ### Changed
 
@@ -17,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 ### Fixed
+- The temperature monitor was starting even when it hadn't been configured.
 
 ### Security